requirements.txt
pycryptodome
mysql-connector-python
